Fix issues, cleanup, refactor code, remove questionable features, add new features. (#2)
Needing to use this GH action but it wasn't working for a few reasons.
In fixing those issues, took the opportunity to clean things up a bit
and add some potentially useful features. Also, removed some features
which didn't seem that valuable to me. Could re-add them if someone came
a long and needed them.